+++ include/dhara/journal.h
